Autodiseqc order/preselection integrated in config instead of popup
I think this is a better solution as you do not get the extra screen and
with this change it also works in the startup wizard.

Note: Maybe needs some finetuning as the rotating polarisation sats were
only selectable when single connection was used. I do not understand why
this was needed.

Co-Authored-By: littlesat <littlesat99@yahoo.com>