Resolve when exit custumize config without changes request to quit
The config.usage.setupShowDefault had a selection with an empty string
causing when exiting the Customumize config that when this option was
selected you always get the popup that you want to exit with changes.
This resolved it. However you still might get it once until you save the
new "no" string.

Co-Authored-By: littlesat <littlesat99@yahoo.com>